# tika-server-old/tika-server-spring/src/main/resources/api/tika-server-openapi.yaml
...
openapi: 3.0.2
...
info:
  contact:
    email: dev@tika.apache.org
  description: The Tika server implements [JAX-RS](http://en.wikipedia.org/wiki/JAX-RS) (Java API for RESTful Web Services) to provide web services according to the Representational State Transfer (REST) architectural style. This facilitates a wide varity oif operations and flexibility with regards to both client and server implementations. The officially supported Tika server implementation is packaged using the OpenAPI [jaxrs-cxf generator](https://openapi-generator.tech/docs/generators/jaxrs-cxf]. This work was tracked through [TIKA-3082](https://issues.apache.org/jira/browse/TIKA-3082).
    <b>N.B.</b> the OpenAPI version always tracks the underlying Tika version to remove uncertainty about which version of Tika is running within the server.
  license:
    name: Apache v2
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Tika JAX-RS Server
  version: 4.0.0-SNAPSHOT
servers:
  - description: Local development server
    url: 'https://localhost:9998'
...
paths:
...
  /rmeta:
    put:
        description: 'Returns an InputStream that can be deserialized as a list of {@link Metadata} objects. The first in the list represents the main document, and the rest represent metadata for the embedded objects. This works recursively through all descendants of the main document, not just the immediate children. The extracted text content is stored with the key {@link RecursiveParserWrapper#TIKA_CONTENT}.'
        operationId: put_rmeta
        responses:
          '200':
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Metadata'
            description: 'Returns a JSONified list of Metadata objects for the container document and all embedded documents.'
          '500':
            description: An error occurred processing the call.
        summary: 'Returns a JSONified list of Metadata objects for the container document and all embedded documents.'
        tags:
        - Recursive Metadata and Content
  /rmeta/{handler}:
    put:
      description: 'Returns an InputStream that can be deserialized as a list of Metadata objects with a specific handler type (xml, text, html, ignore).'
      parameters:
      - description: Handler type for content (xml, text, html, ignore)
        explode: false
        in: path
        name: handler
        required: true
        schema:
          example: 'xml'
          type: string
          enum: [xml, text, html, ignore]
        style: simple
      operationId: put_rmeta_handler
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Metadata'
          description: 'Returns a JSONified list of Metadata objects with specified handler.'
        '500':
          description: An error occurred processing the call.
      summary: 'Returns recursive metadata with specified handler type.'
      tags:
      - Recursive Metadata and Content
  /rmeta/form:
    post:
      description: 'POST a document via multipart/form-data for recursive metadata extraction.'
      operationId: post_rmeta_form
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Metadata'
          description: 'Returns a JSONified list of Metadata objects for the container document and all embedded documents.'
        '500':
          description: An error occurred processing the call.
      summary: 'POST form for recursive metadata extraction.'
      tags:
      - Recursive Metadata and Content
  /rmeta/form/{handler}:
    post:
      description: 'POST a document via multipart/form-data for recursive metadata extraction with specified handler.'
      parameters:
      - description: Handler type for content (xml, text, html, ignore)
        explode: false
        in: path
        name: handler
        required: true
        schema:
          example: 'xml'
          type: string
          enum: [xml, text, html, ignore]
        style: simple
      operationId: post_rmeta_form_handler
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Metadata'
          description: 'Returns a JSONified list of Metadata objects with specified handler.'
        '500':
          description: An error occurred processing the call.
      summary: 'POST form for recursive metadata extraction with handler.'
      tags:
      - Recursive Metadata and Content
...
components:
  schemas:
...
    Metadata:
      description: "The typical Tika multi-valued metadata key-value map."
      type: object
      additionalProperties:
        type: string
...